Updates TcXaeMgmt module installation and loading
Ensures that the script installs or loads TcXaeMgmt module version 6.2.0 or greater, regardless of the PowerShell edition.

This change simplifies the module version check and installation process, ensuring compatibility with a wider range of PowerShell environments. It installs from the PowerShell Gallery and always imports the latest available version that meets the minimum version requirement.